backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysqlclient==2.2.1
pymysql==1.1.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
aiomysql==0.2.0
beautifulsoup4==4.12.2
aiohttp==3.8.5
tenacity==8.2.2
openai==1.61.0
alembic==1.13.1
boto3==1.34.49
botocore==1.34.49

